Renamed ModelService to ApiCommon
