+++ src/app/services/baw-api/mock/api-commonMock.ts
@@ -1,6 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Subject } from "rxjs";
 import { BawApiService } from "../base-api.service";
 
 @Injectable()
-export class MockModelService<T> extends BawApiService {}
+export class MockApiCommon<T> extends BawApiService {}
