+++ src/app/services/baw-api/mock/projectsMockService.ts
@@ -1,12 +1,12 @@
 import { Injectable } from "@angular/core";
 import { Subject } from "rxjs";
 import { Project } from "src/app/models/Project";
-import { MockModelService } from "./modelMockService";
+import { MockApiCommon } from "./api-commonMock";
 
 @Injectable({
   providedIn: "root"
 })
-export class MockProjectsService extends MockModelService<Project> {
+export class MockProjectsService extends MockApiCommon<Project> {
   public getProject() {
     return new Subject();
   }
