+++ src/app/services/baw-api/user.service.ts
@@ -4,12 +4,12 @@ import { Subject } from "rxjs";
 import { ID } from "src/app/interfaces/apiInterfaces";
 import { User, UserInterface } from "src/app/models/User";
 import { AppConfigService } from "../app-config/app-config.service";
-import { ModelService } from "./model.service";
+import { ApiCommon } from "./api-common";
 
 @Injectable({
   providedIn: "root"
 })
-export class UserService extends ModelService<User> {
+export class UserService extends ApiCommon<User> {
   private paths: {
     [key: string]: string;
   };
